# backend/backend/openapi.yaml
openapi: 3.0.3
info:
  title: PetClinic API
  version: 1.0.0
  description: |
    REST API for the Spring PetClinic schema (TypeScript / Express / Sequelize port).

    Interactive docs: `/api/docs`. Raw spec: `/api/docs.json`.

    **Authentication.** Every `/api/*` endpoint requires a Bearer access token
    obtained from `POST /auth/login` or `POST /auth/signup`. Unauthenticated
    requests get `401`. The `/auth/*` endpoints, `/health`, and the docs are
    public.
servers:
  - url: /
    description: Same origin as the running server.
security:
  - bearerAuth: []
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Create an account and receive a token pair
      security: []
      description: |
        Creates a `User` with a single default role (`ROLE_USER`) and returns an
        access + refresh token pair. Use this for public self-service signup.
        For administrative user creation (with arbitrary roles), use `POST /api/users`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Exchange credentials for a token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate a refresh token for a new pair
      security: []
      description: Revokes the presented refresh token and issues a new access + refresh pair. Roles are re-read from the database, so role changes take effect within one refresh.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke a refresh token
      security: []
      description: Revokes the presented refresh token. Access tokens remain valid until they expire; clients are expected to drop them locally.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshInput' }
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the currently authenticated user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me/account:
    patch:
      tags: [Auth]
      summary: Update the current user's account details
      description: 'Only `email` is editable. `username` is the natural primary key and cannot change; `enabled` and `roles` are admin concerns.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountUpdateInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthUser' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/me/password:
    patch:
      tags: [Auth]
      summary: Change the current user's password
      description: Requires the current password. On success, revokes all of the user's refresh tokens and returns a fresh pair.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordChangeInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/password-reset/request:
    post:
      tags: [Auth]
      summary: Request a password reset link
      security: []
      description: |
        Always returns 204 regardless of whether the email is known — this
        avoids leaking which addresses are registered. In dev, the reset link
        is logged to the server console; replace `deliverResetLink` with real
        email delivery for production.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordResetRequestInput' }
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Complete a password reset
      security: []
      description: Consumes a reset token, sets the new password, and revokes all of the user's refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordResetConfirmInput' }
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by `/auth/login`, `/auth/signup`, or `/auth/refresh`.
...
  responses:
    BadRequest:
      description: Malformed request (bad path param, missing required field, validation failure).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid, or expired access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Unique-constraint or foreign-key violation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Owner with id 99 not found
        details:
          type: array
          items: { type: string }
          description: Per-field validation messages (Sequelize validation / unique errors).
        path:
          type: string
          description: Returned for the global "Route not found" 404.
...
    AuthUser:
      type: object
      description: Shape of the current user as returned by the auth endpoints (roles flattened to string names).
      required: [username, enabled, roles]
      properties:
        username: { type: string }
        email:    { type: string, format: email, nullable: true }
        enabled:  { type: boolean }
        roles:
          type: array
          items: { type: string }
          example: [ROLE_USER]
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: 'Short-lived JWT for the `Authorization: Bearer` header.'
        refresh_token:
          type: string
          description: 'Opaque random token; submit to `/auth/refresh` to rotate.'
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Seconds until the access token expires.
    AuthResponse:
      type: object
      description: Returned by signup, login, and password change.
      required: [user, access_token, refresh_token, token_type, expires_in]
      allOf:
        - $ref: '#/components/schemas/TokenPair'
        - type: object
          required: [user]
          properties:
            user: { $ref: '#/components/schemas/AuthUser' }
    SignupInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 20 }
        password: { type: string, format: password, minLength: 8, writeOnly: true }
        email:    { type: string, format: email, nullable: true }
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password, writeOnly: true }
    RefreshInput:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    AccountUpdateInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          nullable: true
          description: Pass `null` to clear the email.
    PasswordChangeInput:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string, format: password, writeOnly: true }
        new_password:     { type: string, format: password, minLength: 8, writeOnly: true }
    PasswordResetRequestInput:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetConfirmInput:
      type: object
      required: [token, new_password]
      properties:
        token:        { type: string }
        new_password: { type: string, format: password, minLength: 8, writeOnly: true }